error-question-system/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
Pillow==10.0.1
PyMySQL==1.1.0
python-json-logger==2.0.7
structlog==23.2.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
drf-spectacular==0.26.5
django-storages==1.14.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.6.0
coverage==7.3.2
flake8==6.1.0
black==23.10.1
isort==5.12.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
whitenoise==6.6.0
flower==2.0.1